requirements.txt
# Core dependencies
openai==1.54.0
python-dotenv==1.0.0

# LangChain ecosystem (compatible versions)
langchain==0.3.7
langchain-openai==0.2.5
langchain-community==0.3.7

# Vector store
chromadb==0.5.20

# Embeddings and reranking
sentence-transformers==3.3.1

# Data handling
pydantic==2.10.2
pydantic-settings==2.6.1

# Utilities
tiktoken==0.8.0
tenacity==9.0.0

# Development
pytest==8.3.4
pytest-asyncio==0.24.0
black==24.10.0
